Avoids checking-in the binary at every commit
