+++ .yarnrc
@@ -1,7 +1,7 @@
-yarn-path "./packages/berry-cli/bin/berry"
+yarn-path "./scripts/run-yarn.js"
 
 init-scope "berry"
 
 plugins:
-  - "./packages/plugin-typescript/bin/@berry/plugin-typescript.js"
-  - "./packages/plugin-pack/bin/@berry/plugin-pack.js"
+  - "./scripts/plugin-pack.js"
+  - "./scripts/plugin-typescript.js"
