--- packages/berry-cli/bin/berry
@@ -1,3 +1,0 @@
-#!/usr/bin/env node
-
-require('./berry.js');
